pyproject.toml
[project]
name = "airflow-dags"
version = "0.1.0"
description = ""
authors = [{ name = "mrpbennett", email = "pbennett.uk@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["psycopg2-binary (>=2.9.10,<3.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
